+++ changelog.md
@@ -9,6 +9,7 @@
   - Fixed %x# variables when multi group optional trigger pattern used
   - Fixed #show and #showprompt breaking telnet ops when split buffer happens
   - Fixed Linux icon not being correctly set in built packages, use mac icns file instead to generate all the correct files, fixes AppImage icons and others
+  - Fixed ansi parsing when remote data is split and appending local text to prevent split ansi codes
 
 ## 0.19.2 2022-08-17